requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bitarray==3.4.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
colorama==0.4.6
cssselect2==0.8.0
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
filelock==3.19.1
Flask==3.1.1
flask-cors==6.0.0
flask-limiter==3.8.0
fonttools==4.58.4
frozenlist==1.6.0
fsspec==2025.9.0
gunicorn==20.1.0
hexbytes==1.3.0
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
multidict==6.4.3
networkx==3.5
numpy==2.2.5
packaging==25.0
pandas==2.2.3
parsimonious==0.10.0
pdfkit==1.0.0
pillow==11.2.1
propcache==0.3.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sympy==1.14.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
weasyprint==65.1
web3==7.11.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.0
zopfli==0.2.3.post1
cryptography==42.0.5
easyocr==1.7.1
torch==2.4.0
transformers==4.44.0
emoji==2.12.1
